+++ packages/react/src/components/Paper/index.ts
@@ -1,0 +1,2 @@
+export { default as Paper } from './Paper';
+export * from './Paper.types';
