+++ packages/react/src/components/Paper/styles.ts
@@ -1,0 +1,8 @@
+export const clickableCss = {
+  cursor: 'pointer',
+  margin: '0.5rem',
+  transition: 'transform 0.1s ease-in-out',
+  ':hover': {
+    transform: 'scale(1.05)',
+  },
+};
