+++ packages/react/src/components/Spacing/spacing.stories.tsx
@@ -2,7 +2,7 @@ import { ComponentStory, ComponentMeta } from '@storybook/react';
 
 import { Button } from '../Button';
 import { FlexBox } from '../FlexBox';
-import Paper from '../Paper';
+import { Paper } from '../Paper';
 import { Text } from '../Text';
 import Spacing from './spacing';
