+++ packages/react/src/index.ts
@@ -9,7 +9,7 @@ export * from './components/Form';
 export { default as TextInput } from './components/TextInput';
 export { default as TextArea } from './components/TextArea';
 export * from './components/LoadingSpinner';
-export { default as Paper } from './components/Paper';
+export * from './components/Paper';
 export { default as Radio } from './components/Radio';
 export { BasicTab, BorderTab } from './components/Tab';
 export { ColumnTable, RowTable, BasicTable } from './components/Table';
